Add files via upload
The task here was to find the association rules between the items using the Apriori Algorithm. Or in other words, we had to find out those items/itemset that customers bought together which helps the owner for store layout/marketing and then show the relative Sales of the data in Tableau Dashboard.